[李瑞瑞]修复跳转场景逻辑。

+++ Assets/Scripts/Comics/ComicsManager.cs
@@ -141,7 +141,7 @@ namespace UIFrame
 
         private List<Sprite> GetSprites()
         {
-            string path = Path.COMICS_PATH + DataManager.Single.LevelIndex.ToString("00");
+            string path = Path.COMICS_PATH + ((int)DataManager.Single.LevelIndex).ToString("00");
             return LoadManager.Single.LoadAll<Sprite>(path).ToList();
         }
         private ComicsItem Move(ComicsParentId id)
