+++ Assets/Scripts/Data/DataManager.cs
@@ -72,7 +72,7 @@ namespace Manager
             }
             else if(JudgeCurrentScene(ConstValue.COMICS_SCENE))
             {
-                return ConstValue.LEVEL_SCENE + "_" + LevelIndex.ToString("00");
+                return ConstValue.LEVEL_SCENE + "_" + ((int)LevelIndex).ToString("00");
             }
             else
             {
